requirements.txt
pygame>=2.5.0
PyYAML>=6.0
pytest>=7.4.0

